requirement.txt
streamlit
deepgram-sdk
requests
python-dotenv
openai